requirements.txt
requests
sqlalchemy
pandas
psycopg2-binary
pytz
apache-airflow-providers-postgres